requirements.txt
numpy
pandas
torch
torchvision
segmentation-models-pytorch
scikit-learn
matplotlib
Pillow
opencv-python
gdal
folium
google-auth-oauthlib
google-auth
google-api-python-client
albumentations
earthengine-api 
rasterio
pytest
openpyxl
